pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adk-extra-services"
dynamic = ["version"]
description = "Additional service implementations for Google ADK"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [{name = "edu010101", email = "eduardo@geoia.tech"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["adk", "google-adk", "s3", "amazon-s3", "artifact-storage", "session-service", "redis", "mongo", "supabase"]
dependencies = [
    "google-adk",
    "google-generativeai>=0.8.5",
    "boto3>=1.28.0",
    "typing-extensions>=4.0.0",
    "motor>=3.3.0",
    "redis>=5.0.0",
    "azure-storage-blob>=12.20.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black",
    "flake8",
    "mypy",
    "isort",
    "python-dotenv>=1.0.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
    "testcontainers>=3.10.0",
]
docs = [
    "sphinx>=7.2.0",
    "furo>=2023.9.10",
    "myst-parser>=2.0.0",
]

[project.urls]
"Homepage" = "https://github.com/edu010101/adk-extra-services"
"Source" = "https://github.com/edu010101/adk-extra-services"
"Documentation" = "https://github.com/edu010101/adk-extra-services#readme"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "adk_extra_services.__version__.__version__"}

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
